Move common code to engine time
