+++ consensus/cryptarchia-engine/Cargo.toml
@@ -10,7 +10,11 @@ version = "0.1.0"
 nomos-utils = { workspace = true, optional = true }
 serde       = { version = "1.0", features = ["derive"], optional = true }
 thiserror   = "1"
+time        = "0.3"
+tokio       = { version = "1", optional = true }
+
 
 [features]
 default = []
 serde   = ["dep:serde", "nomos-utils/serde"]
+tokio   = ["tokio/time"]
