+++ consensus/cryptarchia-engine/src/time.rs
@@ -1,4 +1,8 @@
+use serde::{Deserialize, Serialize};
 use std::ops::Add;
+use std::time::Duration;
+use time::OffsetDateTime;
+use tokio::time::{Interval, MissedTickBehavior};
 
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[derive(Clone, Debug, Eq, PartialEq, Copy, Hash, PartialOrd, Ord)]
@@ -15,6 +19,20 @@ impl Slot {
 
     pub fn genesis() -> Self {
         Self(0)
+    }
+
+    pub fn current_from_offset_and_config(
+        offset_date_time: OffsetDateTime,
+        slot_config: SlotConfig,
+    ) -> Self {
+        // TODO: leap seconds / weird time stuff
+        let since_start = offset_date_time - slot_config.chain_start_time;
+        if since_start.is_negative() {
+            // current slot is behind the start time, so return default 0
+            Slot::genesis()
+        } else {
+            Slot::from(since_start.whole_seconds() as u64 / slot_config.slot_duration.as_secs())
+        }
     }
 }
 
@@ -59,7 +77,7 @@ impl Add<u32> for Epoch {
 }
 
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Debug, PartialEq)]
+#[derive(Copy, Clone, Debug, PartialEq)]
 pub struct EpochConfig {
     // The stake distribution is always taken at the beginning of the previous epoch.
     // This parameters controls how many slots to wait for it to be stabilized
@@ -83,5 +101,44 @@ impl EpochConfig {
 
     pub fn epoch(&self, slot: Slot, base_period_length: u64) -> Epoch {
         ((u64::from(slot) / base_period_length) as u32).into()
+    }
+}
+
+#[derive(Copy, Clone, Debug)]
+pub struct SlotConfig {
+    pub slot_duration: Duration,
+    /// Start of the first epoch
+    pub chain_start_time: OffsetDateTime,
+}
+
+#[cfg(feature = "tokio")]
+#[derive(Clone, Debug)]
+pub struct SlotTimer {
+    config: SlotConfig,
+}
+
+#[cfg(feature = "tokio")]
+impl SlotTimer {
+    pub fn new(config: SlotConfig) -> Self {
+        SlotTimer { config }
+    }
+
+    pub fn current_slot(&self) -> Slot {
+        Slot::current_from_offset_and_config(OffsetDateTime::now_utc(), self.config)
+    }
+
+    /// Ticks at the start of each slot, starting from the next slot
+    pub fn slot_interval(&self, now: OffsetDateTime) -> Interval {
+        let slot_duration = self.config.slot_duration;
+        let next_slot_start = self.config.chain_start_time
+            + slot_duration * u64::from(self.current_slot() + 1) as u32;
+        let delay = next_slot_start - now;
+        let mut interval = tokio::time::interval_at(
+            tokio::time::Instant::now()
+                + Duration::try_from(delay).expect("could not set slot timer duration"),
+            slot_duration,
+        );
+        interval.set_missed_tick_behavior(MissedTickBehavior::Skip);
+        interval
     }
 }
